Set TERM_PROGRAM and TERM_PROGRAM_VERSION environment variables.
WezTerm claims this is an emerging de-facto standard for terminal emulator identification:
https://github.com/wez/wezterm/blob/a103b6d97adcc439e27cec0d3bb2c0f72a79095d/config/src/config.rs#L1526-L1529

One example of usage in the wild is neovim doing capability detection:
https://github.com/neovim/neovim/blob/f050aaabbb31b58ebac519a6f7014610f2b119f0/src/nvim/tui/tui.c#L206-L211

Ghostty now reports this:

$echo $TERM_PROGRAM
ghostty

$echo $TERM_PROGRAM_VERSION
0.1.0-main+cfcf065

I think it's really nice that the commit hash is included, as users can provide this in issue reports. WezTerm does the same.

I use these variables in my tui library in addition to $TERM and $COLORTERM for capability detection, which is what motivated this PR.